Rename Grid to Masonry
The current Grid implementation is actually a Masonry.
We'll be adding a proper Grid implementation with this branch.
