+++ src/components/Lists/List.vue
@@ -1,5 +1,5 @@
 <template lang="html">
-  <div :class="['list', viewClass, { unique: unique && view !== 'grid' }]">
+  <div :class="['list', viewClass, { unique: unique && view !== 'masonry' }]">
     <header>
       <span class="list-name">
         <i
@@ -14,13 +14,13 @@
     </header>
 
     <div
-      v-if="view === 'grid'"
-      :class="`game-grid game-grid-${listIndex}`"
+      v-if="view === 'masonry'"
+      :class="`game-masonry game-masonry-${listIndex}`"
     >
       <component
         v-for="game in gameList"
         :is="gameCardComponent"
-        :key="`grid-${game}`"
+        :key="`masonry-${game}`"
         :id="game"
         :game-id="game"
         :list-id="listIndex"
@@ -40,7 +40,7 @@
       <component
         v-for="game in sortedGames"
         :is="gameCardComponent"
-        :key="`grid-${game}`"
+        :key="`masonry-${game}`"
         :id="game"
         :game-id="game"
         :list-id="listIndex"
@@ -56,7 +56,7 @@ import draggable from 'vuedraggable';
 import Masonry from 'masonry-layout';
 import ListSettingsModal from '@/components/Lists/ListSettingsModal';
 import GameCardDefault from '@/components/GameCards/GameCardDefault';
-import GameCardGrid from '@/components/GameCards/GameCardGrid';
+import GameCardMasonry from '@/components/GameCards/GameCardMasonry';
 import GameCardWide from '@/components/GameCards/GameCardWide';
 import GameCardText from '@/components/GameCards/GameCardText';
 import AddGameModal from '@/components/Lists/AddGameModal';
@@ -67,7 +67,7 @@ export default {
 
   components: {
     GameCardDefault,
-    GameCardGrid,
+    GameCardMasonry,
     GameCardWide,
     GameCardText,
     AddGameModal,
@@ -103,7 +103,7 @@ export default {
       },
       gameCardComponents: {
         single: 'GameCardDefault',
-        grid: 'GameCardGrid',
+        masonry: 'GameCardMasonry',
         wide: 'GameCardWide',
         text: 'GameCardText',
       },
@@ -192,36 +192,36 @@ export default {
 
   watch: {
     view() {
-      this.initGrid();
+      this.initMasonry();
 
       setTimeout(() => {
-        this.initGrid();
+        this.initMasonry();
       }, 500);
     },
 
     gameList() {
-      this.initGrid();
+      this.initMasonry();
 
       setTimeout(() => {
-        this.initGrid();
+        this.initMasonry();
       }, 500);
     },
   },
 
   mounted() {
-    this.initGrid();
+    this.initMasonry();
 
     setTimeout(() => {
-      this.initGrid();
+      this.initMasonry();
     }, 500);
   },
 
   methods: {
-    initGrid() {
-      if (this.view === 'grid') {
+    initMasonry() {
+      if (this.view === 'masonry') {
         this.$nextTick(() => {
           // eslint-disable-next-line
-            this.masonry = new Masonry(`.game-grid-${this.listIndex}`, {
+            this.masonry = new Masonry(`.game-masonry-${this.listIndex}`, {
             itemSelector: '.game-card',
             gutter: 4,
           });
@@ -319,7 +319,7 @@ export default {
     padding: $gp;
   }
 
-  .game-grid {
+  .game-masonry {
     height: 100%;
     display: flex;
     align-items: center;
