+++ src/i18n/ar.json
@@ -105,6 +105,7 @@
       "single": "افتراضي",
       "wide": "المدمج",
       "text": "نص فقط",
+      "masonry": "Masonry",
       "grid": "شبكة"
     }
   },
