+++ src/i18n/cs.json
@@ -105,6 +105,7 @@
       "single": "Výchozí",
       "wide": "Kompaktní",
       "text": "Pouze text",
+      "masonry": "Masonry",
       "grid": "Mřížka"
     }
   },
