+++ src/i18n/es.json
@@ -105,6 +105,7 @@
       "single": "Defecto",
       "wide": "Compacto",
       "text": "Solo texto",
+      "masonry": "Masonry",
       "grid": "Cuadrícula"
     }
   },
