+++ src/i18n/fr.json
@@ -105,6 +105,7 @@
       "single": "Défaut",
       "wide": "Compact",
       "text": "Texte seulement",
+      "masonry": "Masonry",
       "grid": "la grille"
     }
   },
