+++ src/i18n/it.json
@@ -105,6 +105,7 @@
       "single": "Predefinito",
       "wide": "Compatto",
       "text": "Solo testo",
+      "masonry": "Masonry",
       "grid": "Griglia"
     }
   },
