+++ src/i18n/pl.json
@@ -105,6 +105,7 @@
       "single": "Domyślna",
       "wide": "Kompaktowy",
       "text": "Tylko tekst",
+      "masonry": "Masonry",
       "grid": "Krata"
     }
   },
